submodule refactoring

+++ src/main/java/dev/n1t/authentication/DTO/UserDTO.java
@@ -1,5 +1,5 @@
 package dev.n1t.authentication.DTO;
 
-public record UserDTO(int id, String firstname, String lastname, String email, String password) {
+public record UserDTO(Long id, String firstname, String lastname, String email, String password) {
 
 }
