+++ src/main/java/dev/n1t/authentication/DTO/UserWithTokenDTO.java
@@ -1,6 +1,6 @@
 package dev.n1t.authentication.DTO;
 
-public record UserWithTokenDTO( Integer id, String firstname, String lastname, String email, String AccessToken, String RefreshToken) {
+public record UserWithTokenDTO( Long id, String firstname, String lastname, String email, String AccessToken, String RefreshToken) {
 
 
 }
