+++ src/main/java/dev/n1t/authentication/repositories/RoleRepository.java
@@ -1,11 +1,12 @@
 package dev.n1t.authentication.repositories;
 
-import dev.n1t.authentication.models.Role;
+
+import dev.n1t.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Optional<Role> getRoleById(Integer id);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Optional<Role> getRoleById(Long id);
     Optional<Role> getRoleByRoleName(String roleName);;
 }
